Did logs and got filters to work

+++ ApplicationCmds/phish/config.js
@@ -1,7 +1,7 @@
 const { MessageEmbed } = require('discord.js');
 
 module.exports = {
-    name: "config",
+    name: "configure",
     description: "Set the configuration for your server",
     options: [
         {
@@ -10,13 +10,35 @@ module.exports = {
             type: 1,
         },
         {
-            name: "create",
+            name: "bypass",
+            description: "Create or remove links that will bypass the filters",
+            type: 1,
+            options: [
+                {
+                    name: "create",
+                    description: "Create a bypass link",
+                    type: 3,
+                },
+                {
+                    name: "remove",
+                    description: "Remove a bypass link",
+                    type: 3,
+                },
+            ]
+        },
+        {
+            name: "set",
             description: "Set configuration for your server",
             type: 1,
             options: [
                 {
                     name: "delete",
                     description: "If the detected phishing links should be deleted.",
+                    type: 5,
+                },
+                {
+                    name: "youtube-filter",
+                    description: "Scan and filter \"free nitro generator\" youtube videos.",
                     type: 5,
                 },
                 {
@@ -52,13 +74,17 @@ module.exports = {
             type: 1,
             options: [
                 {
-                    name: "reset",
+                    name: "configurations",
                     description: "Reset the server configurations back to it's original state.",
                     type: 3,
                     choices: [
                         {
                             name: "Delete",
                             value: "delete"
+                        },
+                        {
+                            name: "Youtube-filter",
+                            value: "youtube"
                         },
                         {
                             name: "Log",
@@ -77,15 +103,39 @@ module.exports = {
             ]
         },
     ],
+    permissions: "MANAGE_GUILD",
     category: "phish",
     run: async(interaction,  client) => {
-        if(interaction.options._subcommand === "get") {
-            let configEmbed = new MessageEmbed()
-                .setTitle(`The current configuration for \`${interaction.guild.name}\``)
-                .setColor(0x00AE86)
-                .setDescription(`\`\`\`diff\n${await client.db.fetch(`${interaction.guild.id}.config.delete`) ? "+" : "-"} Delete Links: ${await client.db.fetch(`${interaction.guild.id}.config.delete`) ? "On" : "Off"}\n${await client.db.fetch(`${interaction.guild.id}.config.log.id`) ? "+" : "-"} Logging: ${await client.db.fetch(`${interaction.guild.id}.config.log.id`) ? "On" : "Off"}\`\`\``)
-            interaction.reply({embeds: [configEmbed]});
+        console.log(interaction.options)
+        switch(interaction.options._subcommand) {
+            case "get":
+                let configEmbed = new MessageEmbed()
+                    .setTitle(`The current configuration for \`${interaction.guild.name}\``)
+                    .setColor(0x00AE86)
+                    .setDescription(`\`\`\`diff\n++++ Server Configurations ++++\n${await client.db.fetch(`${interaction.guild.id}.config.delete`) ? "+" : "-"} Delete Links: ${await client.db.fetch(`${interaction.guild.id}.config.delete`) ? "On" : "Off"}\n${await client.db.fetch(`${interaction.guild.id}.config.youtube`) ? "+" : "-"} Youtube Filter: ${await client.db.fetch(`${interaction.guild.id}.config.youtube`) ? "On" : "Off"}\n${await client.db.fetch(`${interaction.guild.id}.config.log.id`) ? "+" : "-"} Logging: ${await client.db.fetch(`${interaction.guild.id}.config.log.id`) ? "On" : "Off"}\n  Actions:\n${await client.db.fetch(`${interaction.guild.id}.config.ban`) ? "+" : "-"} Ban: ${await client.db.fetch(`${interaction.guild.id}.config.ban`) ? "On" : "off"}\n${await client.db.fetch(`${interaction.guild.id}.config.kick`) ? "+" : "-"} Kick: ${await client.db.fetch(`${interaction.guild.id}.config.kick`) ? "On" : "Off"}\n${await client.db.fetch(`${interaction.guild.id}.config.timeout`) ? "+" : "-"} Timeout: ${await client.db.fetch(`${interaction.guild.id}.config.guild`) ? "On" : "Off"}\`\`\``)
+                interaction.reply({embeds: [configEmbed]});
+                break
+
+            case "bypass":
+                if(!interaction.options._hoistedOptions[0]) return interaction.reply({content: "<:3595failed:926715200172867624> You must specify a option!", ephemeral: true})
+                if(interaction.options._hoistedOptions.length > 1) return interaction.reply({content: "<:3595failed:926715200172867624> You can only specify one option!", ephemeral: true})
+                const string = interaction.options._hoistedOptions[0].value;
+                switch(interaction.options._hoistedOptions[0].name) {
+
+                    case "create":
+                        if(!interaction.options._hoistedOptions[0]) return interaction.reply({content: "<:3595failed:926715200172867624> You must specify a option!", ephemeral: true})
+                        if(interaction.options._hoistedOptions.length > 1) return interaction.reply({content: "<:3595failed:926715200172867624> You can only specify one option!", ephemeral: true})
+                        interaction.reply({content: 'This command is still being worked on!', ephemeral: true})
+                        break
+                    case "remove":
+                        interaction.reply({content: 'This command is still being worked on!', ephemeral: true})
+                        break
+
+                }
+
+                break
         }
+
         const value = interaction.options._hoistedOptions
         if(value[0]) {
             if(!value[0]) return interaction.reply({content: "<:3595failed:926715200172867624> You must specify a option!", ephemeral: true})
@@ -96,6 +146,11 @@ module.exports = {
                     if(await client.db.fetch(`${interaction.guild.id}.config.delete`)) return interaction.reply({content: "<:3595failed:926715200172867624> Phishing links are already being deleted!", ephemeral: true})
                     client.db.set(`${interaction.guild.id}.config.delete`, interaction.options._hoistedOptions[0].value)
                     interaction.reply({content: "<:9294passed:926715199950561341> The detected phishing or malicious links will now be deleted.", ephemeral: true})
+                    break
+                case "youtube-filter":
+                    if(await client.db.fetch(`${interaction.guild.id}.config.youtube`)) return interaction.reply({content: "<:3595failed:926715200172867624> Youtube links are already being filtered.", ephemeral: true})
+                    client.db.set(`${interaction.guild.id}.config.youtube`, interaction.options._hoistedOptions[0].value)
+                    interaction.reply({content: "<:9294passed:926715199950561341> Successfully configured, I will now filter youtube videos that get sent.", ephemeral: true})
                     break
                 case "action":
 
@@ -156,7 +211,7 @@ module.exports = {
 
                     break;
 
-                case "reset":
+                case "configurations":
 
                     switch(value[0].value) {
                         case "delete":
@@ -164,10 +219,12 @@ module.exports = {
                             interaction.reply({content: "<:9294passed:926715199950561341> The configuration has been reset!", ephemeral: true})
                             break
                         case "log":
-                            const channelFetch = await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`)).fetchWebhooks();
-                            const filter = channelFetch.filter(webhook => webhook.owner.id === client.user.id && webhook.name === 'Phish Logger')
-                            if (filter.size === 0) return
-                            for (let [id, webhook] of filter) await webhook.delete();
+                            if(await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`))) {
+                                const loggerWebhook = await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`)).fetchWebhooks()
+                                const deleteFilter = loggerWebhook.filter(webhook => webhook.owner.id === client.user.id && webhook.name === 'Phish Logger')
+                                if (deleteFilter.size === 0) return
+                                for (let [id, webhook] of deleteFilter) await webhook.delete();
+                            }
                             client.db.delete(`${interaction.guild.id}.config.log.channelId`)
                             client.db.delete(`${interaction.guild.id}.config.log.id`)
                             client.db.delete(`${interaction.guild.id}.config.log.token`)
@@ -181,19 +238,30 @@ module.exports = {
                             client.db.delete(`${interaction.guild.id}.config.timeout`)
                             interaction.reply({content: "<:9294passed:926715199950561341> The configuration has been reset!", ephemeral: true})
                             break
+                        case "youtube":
+                            client.db.delete(`${interaction.guild.id}.config.youtube`)
+                            interaction.reply({content: "<:9294passed:926715199950561341> The configuration has been reset!", ephemeral: true})
+                            break
                         case "all":
-                            const loggerWebhook = await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`)).fetchWebhooks();
-                            const deleteFilter = loggerWebhook.filter(webhook => webhook.owner.id === client.user.id && webhook.name === 'Phish Logger')
-                            if (deleteFilter.size === 0) return
-                            for (let [id, webhook] of deleteFilter) await webhook.delete();
-                            client.db.delete(`${interaction.guild.id}.config.delete`)
-                            client.db.delete(`${interaction.guild.id}.config.log.channelId`)
-                            client.db.delete(`${interaction.guild.id}.config.log.id`)
-                            client.db.delete(`${interaction.guild.id}.config.log.token`)
-                            client.db.delete(`${interaction.guild.id}.config.ban`)
-                            client.db.delete(`${interaction.guild.id}.config.kick`)
-                            client.db.delete(`${interaction.guild.id}.config.timeout`)
-                            interaction.reply({content: "<:9294passed:926715199950561341> All configurations have been reset!", ephemeral: true})
+
+                            if(await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`))) {
+                                const loggerWebhook = await client.channels.cache.get(await client.db.fetch(`${interaction.guild.id}.config.log.channelId`)).fetchWebhooks()
+                                const deleteFilter = loggerWebhook.filter(webhook => webhook.owner.id === client.user.id && webhook.name === 'Phish Logger')
+                                if (deleteFilter.size === 0) return
+                                for (let [id, webhook] of deleteFilter) await webhook.delete();
+                            }
+                                client.db.delete(`${interaction.guild.id}.config.delete`)
+                                client.db.delete(`${interaction.guild.id}.config.log.channelId`)
+                                client.db.delete(`${interaction.guild.id}.config.log.id`)
+                                client.db.delete(`${interaction.guild.id}.config.log.token`)
+                                client.db.delete(`${interaction.guild.id}.config.youtube`)
+                                client.db.delete(`${interaction.guild.id}.config.ban`)
+                                client.db.delete(`${interaction.guild.id}.config.kick`)
+                                client.db.delete(`${interaction.guild.id}.config.timeout`)
+                                interaction.reply({content: "<:9294passed:926715199950561341> All configurations have been reset!", ephemeral: true})
+
+
+
                             break
                     }
